tools/auto-flash/pyproject.toml
[tool.poetry]
name = "auto-flash"
version = "0.1.0"
description = ""
authors = ["Michel Meyer <meyer_michel@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.5"
dataclasses-json = "^0.5.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
